clear menu data for first time login only
