+++ src/app/login/login.component.ts
@@ -56,7 +56,7 @@ export class LoginComponent implements OnInit, OnDestroy {
   public isForgotPasswordInSuccess$: Observable<boolean>;
   public isResetPasswordInSuccess$: Observable<boolean>;
   public signupVerifyForm: FormGroup;
-  public isLoginWithPasswordSuccessNotVerified$:Observable<boolean>;
+  public isLoginWithPasswordSuccessNotVerified$: Observable<boolean>;
 
   public showForgotPassword: boolean = false;
   public forgotStep: number = 0;
@@ -70,13 +70,13 @@ export class LoginComponent implements OnInit, OnDestroy {
 
   // tslint:disable-next-line:no-empty
   constructor(private _fb: FormBuilder,
-    private store: Store<AppState>,
-    private router: Router,
-    private loginAction: LoginActions,
-    private authService: AuthService,
-    @Inject(DOCUMENT) private document: Document,
-    private _toaster: ToasterService,
-    private _authService: AuthenticationService
+              private store: Store<AppState>,
+              private router: Router,
+              private loginAction: LoginActions,
+              private authService: AuthService,
+              @Inject(DOCUMENT) private document: Document,
+              private _toaster: ToasterService,
+              private _authService: AuthenticationService
   ) {
     this.urlPath = isElectron ? '' : AppUrl + APP_FOLDER;
     this.isLoginWithEmailInProcess$ = store.select(state => {
@@ -128,7 +128,7 @@ export class LoginComponent implements OnInit, OnDestroy {
     this.isSocialLogoutAttempted$ = this.store.select(p => p.login.isSocialLogoutAttempted).pipe(takeUntil(this.destroyed$));
 
     contriesWithCodes.map(c => {
-      this.countryCodeList.push({ value: c.countryName, label: c.value });
+      this.countryCodeList.push({value: c.countryName, label: c.value});
     });
     this.userLoginState$ = this.store.select(p => p.session.userLoginState);
     this.userDetails$ = this.store.select(p => p.session.user);
@@ -139,9 +139,9 @@ export class LoginComponent implements OnInit, OnDestroy {
   // tslint:disable-next-line:no-empty
   public ngOnInit() {
 
-    this.emailVerifyModal.config = { backdrop: 'static'};
-    this.twoWayAuthModal.config = { backdrop: 'static'};
-    this.mobileVerifyModal.config = { backdrop: 'static'};
+    this.emailVerifyModal.config = {backdrop: 'static'};
+    this.twoWayAuthModal.config = {backdrop: 'static'};
+    this.mobileVerifyModal.config = {backdrop: 'static'};
 
     this.getElectronAppVersion();
     this.document.body.classList.remove('unresponsive');
@@ -176,7 +176,7 @@ export class LoginComponent implements OnInit, OnDestroy {
       email: ['', [Validators.required, Validators.email]],
       verificationCode: ['', Validators.required]
     });
-    this.setCountryCode({ value: 'India', label: 'India' });
+    this.setCountryCode({value: 'India', label: 'India'});
 
     // get user object when google auth is complete
     if (!Configuration.isElectron) {
@@ -238,11 +238,11 @@ export class LoginComponent implements OnInit, OnDestroy {
         return this.resetTwoWayAuthModal();
       }
     });
-    this.isLoginWithPasswordSuccessNotVerified$.subscribe(res=> {
-      if(res) {
+    this.isLoginWithPasswordSuccessNotVerified$.subscribe(res => {
+      if (res) {
         console.log('isLoginWithPasswordSuccessNotVerified', res);
       }
-    })
+    });
   }
 
   public showEmailModal() {
@@ -420,7 +420,7 @@ export class LoginComponent implements OnInit, OnDestroy {
   }
 
   public forgotPassword(userId) {
-    this.resetPasswordForm.patchValue({ uniqueKey: userId });
+    this.resetPasswordForm.patchValue({uniqueKey: userId});
     this.userUniqueKey = userId;
     this.store.dispatch(this.loginAction.forgotPasswordRequest(userId));
   }
