+++ src/app/services/general.service.ts
@@ -2,11 +2,13 @@ import { Injectable } from '@angular/core';
 import { UserDetails } from '../models/api-models/loginModels';
 import { BehaviorSubject, Subject } from 'rxjs';
 import { eventsConst } from 'app/shared/header/components/eventsConst';
+import { IUlist } from '../models/interfaces/ulist.interface';
 
 @Injectable()
 export class GeneralService {
 
   public talkToSalesModal: BehaviorSubject<boolean> = new BehaviorSubject(false);
+  public menuClickedFromOutSideHeader: Subject<IUlist> = new Subject();
   public isCurrencyPipeLoaded: boolean = false;
 
   get user(): UserDetails {
